Include much more k8s debug info when run monitoring fails (#13096)
Summary:
Like we already have with the EcsRunLauncher, when a k8s job crashes or
fails to start, include:
- the last 10 logs from the pod
- the status of the containers in the pod
- any warning events in the pod

## Summary & Motivation

## How I Tested These Changes

+++ integration_tests/test_suites/k8s-test-suite/tests/test_utils.py
@@ -7,29 +7,36 @@ from dagster_k8s.client import DagsterK8sError, DagsterKubernetesClient, WaitFor
 pytest_plugins = ["dagster_k8s_test_infra.helm"]
 
 
-def construct_pod_spec(name, cmd):
+def construct_pod_spec(name, cmd, image="busybox", container_kwargs=None):
     return kubernetes.client.V1PodSpec(
         restart_policy="Never",
         containers=[
-            kubernetes.client.V1Container(name=name, image="busybox", args=["/bin/sh", "-c", cmd])
+            kubernetes.client.V1Container(
+                name=name,
+                image=image,
+                args=["/bin/sh", "-c", cmd],
+                **(container_kwargs or {}),
+            )
         ],
     )
 
 
-def construct_pod_manifest(name, cmd):
+def construct_pod_manifest(name, cmd, image="busybox", container_kwargs=None):
     return kubernetes.client.V1Pod(
         metadata=kubernetes.client.V1ObjectMeta(name=name),
-        spec=construct_pod_spec(name, cmd),
+        spec=construct_pod_spec(name, cmd, image=image, container_kwargs=container_kwargs),
     )
 
 
-def construct_job_manifest(name, cmd):
+def construct_job_manifest(name, cmd, image="busybox", container_kwargs=None):
     return kubernetes.client.V1Job(
         api_version="batch/v1",
         kind="Job",
         metadata=kubernetes.client.V1ObjectMeta(name=name),
         spec=kubernetes.client.V1JobSpec(
-            template=kubernetes.client.V1PodTemplateSpec(spec=construct_pod_spec(name, cmd)),
+            template=kubernetes.client.V1PodTemplateSpec(
+                spec=construct_pod_spec(name, cmd, image=image, container_kwargs=container_kwargs)
+            ),
         ),
     )
 
@@ -91,7 +98,264 @@ def test_wait_for_pod(cluster_provider, namespace):
 
 
 @pytest.mark.default
-def test_wait_for_job(cluster_provider, namespace):
+def test_pod_debug_info_failure(cluster_provider, namespace, should_cleanup):
+    time.sleep(5)
+    api_client = DagsterKubernetesClient.production_client()
+
+    try:
+        # test case where the pod can't be scheduled due to an unworkable resource request
+        api_client.batch_api.create_namespaced_job(
+            body=construct_job_manifest(
+                "resourcelimit",
+                'echo "hello world"',
+                container_kwargs={
+                    "resources": kubernetes.client.V1ResourceRequirements(
+                        requests={"memory": "4000000M"}, limits={"memory": "4000000M"}
+                    ),
+                },
+            ),
+            namespace=namespace,
+        )
+        api_client.wait_for_job("resourcelimit", namespace=namespace)
+
+        pod_names = api_client.get_pod_names_in_job("resourcelimit", namespace=namespace)
+
+        pod_debug_info = api_client.get_pod_debug_info(
+            pod_names[0], namespace=namespace, container_name="resourcelimit"
+        )
+
+        print(str(pod_debug_info))  # noqa
+
+        assert pod_debug_info.startswith(
+            f"""Debug information for pod {pod_names[0]}:
+Pod status: Pending
+Warning events for pod:
+FailedScheduling: 0/1 nodes are available: 1 Insufficient memory."""
+        )
+
+        # Test case where the pod is still running
+        api_client.batch_api.create_namespaced_job(
+            body=construct_job_manifest(
+                "waitforever", 'echo "time for sleep"; sleep 3600; echo "hello world"'
+            ),
+            namespace=namespace,
+        )
+        api_client.wait_for_job("waitforever", namespace=namespace)
+
+        start_time = time.time()
+        while True:
+            if time.time() - start_time > 60:
+                raise Exception("No error state after 60 seconds")
+
+            pod_names = api_client.get_pod_names_in_job("waitforever", namespace=namespace)
+            if pod_names:
+                pod_debug_info = api_client.get_pod_debug_info(
+                    pod_names[0], namespace=namespace, container_name="waitforever"
+                )
+                if "time for sleep" in pod_debug_info:
+                    break
+
+            time.sleep(5)
+
+        print(str(pod_debug_info))  # noqa
+
+        assert pod_debug_info.startswith(
+            f"""Debug information for pod {pod_names[0]}:
+Pod status: Running
+Container 'waitforever' status: Ready
+Last 25 log lines:"""
+        )
+
+        assert "No warning events for pod." in pod_debug_info
+
+        # Test case where logs have an exec format error (simulated)
+        api_client.batch_api.create_namespaced_job(
+            body=construct_job_manifest(
+                "execformaterror", 'echo "exec /usr/local/bin/dagster: exec format error"; exit 1'
+            ),
+            namespace=namespace,
+        )
+
+        with pytest.raises(
+            DagsterK8sError,
+            match="Encountered failed job pods for job execformaterror with status:",
+        ):
+            api_client.wait_for_job_success("execformaterror", namespace=namespace)
+
+        pod_names = api_client.get_pod_names_in_job("execformaterror", namespace=namespace)
+
+        pod_debug_info = api_client.get_pod_debug_info(
+            pod_names[0], namespace=namespace, container_name="execformaterror"
+        )
+
+        print(str(pod_debug_info))  # noqa
+
+        assert pod_debug_info.startswith(
+            f"""Debug information for pod {pod_names[0]}:
+Pod status: Failed
+Container 'execformaterror' status: Terminated with exit code 1: Error
+Pod logs contained `exec format error`, which usually means that your Docker image was built using the wrong architecture.
+Try rebuilding your docker image with the `--platform linux/amd64` flag set.
+Last 25 log lines:
+"""
+        )
+
+        assert "exec /usr/local/bin/dagster: exec format error" in pod_debug_info
+        assert "No warning events for pod." in pod_debug_info
+
+        # Test case where a non-existant secret is used
+        bad_secret_config = [
+            kubernetes.client.V1EnvFromSource(
+                secret_ref=kubernetes.client.V1SecretEnvSource(name="missing-secret")
+            )
+        ]
+
+        api_client.batch_api.create_namespaced_job(
+            body=construct_job_manifest(
+                "missingsecret",
+                'echo "hello world"',
+                container_kwargs={"env_from": bad_secret_config},
+            ),
+            namespace=namespace,
+        )
+
+        api_client.wait_for_job("missingsecret", namespace=namespace)
+
+        start_time = time.time()
+        while True:
+            if time.time() - start_time > 60:
+                raise Exception("No error state after 60 seconds")
+
+            pod_names = api_client.get_pod_names_in_job("missingsecret", namespace=namespace)
+
+            if pod_names:
+                pod_debug_info = api_client.get_pod_debug_info(
+                    pod_names[0], namespace=namespace, container_name="missingsecret"
+                )
+                if "CreateContainerConfigError" in pod_debug_info:
+                    break
+
+            time.sleep(5)
+
+        pod_debug_info = api_client.get_pod_debug_info(
+            pod_names[0], namespace=namespace, container_name="missingsecret"
+        )
+
+        print(str(pod_debug_info))  # noqa
+
+        assert pod_debug_info.startswith(
+            f"""Debug information for pod {pod_names[0]}:
+Pod status: Pending
+Container 'missingsecret' status: Waiting: CreateContainerConfigError: secret "missing-secret" not found
+Warning events for pod:"""
+        )
+
+        # Test case where an unpullable image is used
+        api_client.batch_api.create_namespaced_job(
+            body=construct_job_manifest("pullfail", 'echo "should not reach"', image="fakeimage"),
+            namespace=namespace,
+        )
+
+        api_client.wait_for_job("pullfail", namespace=namespace)
+
+        start_time = time.time()
+        while True:
+            if time.time() - start_time > 60:
+                raise Exception("No error state after 60 seconds")
+
+            pod_names = api_client.get_pod_names_in_job("pullfail", namespace=namespace)
+
+            if pod_names:
+                pod_debug_info = api_client.get_pod_debug_info(
+                    pod_names[0], namespace=namespace, container_name="pullfail"
+                )
+                if "ImagePullBackOff" in pod_debug_info:
+                    break
+
+            time.sleep(5)
+
+        print(str(pod_debug_info))  # noqa
+        assert pod_debug_info.startswith(
+            f"""Debug information for pod {pod_names[0]}:
+Pod status: Pending
+Container 'pullfail' status: Waiting: ErrImagePull: rpc error: code = Unknown desc = failed to pull and unpack image "docker.io/library/fakeimage:latest": failed to resolve reference "docker.io/library/fakeimage:latest": pull access denied, repository does not exist or may require authorization: server message: insufficient_scope: authorization failed
+Warning events for pod:"""
+        )
+        assert "Failed: Error: ErrImagePull" in pod_debug_info
+        assert "Failed: Error: ImagePullBackOff" in pod_debug_info
+
+        # Test case where the pod unexpectedly terminates
+        api_client.batch_api.create_namespaced_job(
+            body=construct_job_manifest("fail", 'echo "whoops!"; exit 1'),
+            namespace=namespace,
+        )
+
+        with pytest.raises(
+            DagsterK8sError,
+            match="Encountered failed job pods for job fail with status:",
+        ):
+            api_client.wait_for_job_success("fail", namespace=namespace)
+
+        pod_names = api_client.get_pod_names_in_job("fail", namespace=namespace)
+
+        pod_debug_info = api_client.get_pod_debug_info(
+            pod_names[0], namespace=namespace, container_name="fail"
+        )
+
+        print(pod_debug_info)  # noqa
+
+        assert pod_debug_info.startswith(
+            f"""Debug information for pod {pod_names[0]}:
+Pod status: Failed
+Container 'fail' status: Terminated with exit code 1: Error
+Last 25 log lines:"""
+        )
+        assert " whoops!\n" in pod_debug_info
+        assert pod_debug_info.endswith("No warning events for pod.")
+
+        # Test case where the pod completes successfully
+        api_client.batch_api.create_namespaced_job(
+            body=construct_job_manifest("sayhi1", 'echo "hello world"'), namespace=namespace
+        )
+        api_client.wait_for_job_success("sayhi1", namespace=namespace)
+
+        pod_names = api_client.get_pod_names_in_job("sayhi1", namespace=namespace)
+
+        pod_debug_info = api_client.get_pod_debug_info(
+            pod_names[0], namespace=namespace, container_name="sayhi1"
+        )
+        print(pod_debug_info)  # noqa
+
+        assert pod_debug_info.startswith(
+            f"""Debug information for pod {pod_names[0]}:
+Pod status: Succeeded
+Container 'sayhi1' status: Terminated with exit code 0: Completed
+Last 25 log lines:"""
+        )
+        assert "hello world" in pod_debug_info
+        assert "No warning events for pod." in pod_debug_info
+
+    finally:
+        if should_cleanup:
+            for job in [
+                "fail",
+                "pullfail",
+                "missingsecret",
+                "sayhi1",
+                "waitforever",
+                "execformaterror",
+                "resourcelimit",
+            ]:
+                try:
+                    api_client.batch_api.delete_namespaced_job(
+                        job, namespace=namespace, propagation_policy="Foreground"
+                    )
+                except kubernetes.client.rest.ApiException:
+                    pass
+
+
+@pytest.mark.default
+def test_wait_for_job(cluster_provider, namespace, should_cleanup):
     # Without this sleep, we get the following error on kind:
     # HTTP response body:
     # {"kind":"Status","apiVersion":"v1","metadata":{},"status":"Failure","message":"No API
@@ -129,10 +393,11 @@ def test_wait_for_job(cluster_provider, namespace):
             api_client.wait_for_job_success("fail", namespace=namespace)
 
     finally:
-        for job in ["sayhi1", "sayhi2", "fail"]:
-            try:
-                api_client.batch_api.delete_namespaced_job(
-                    job, namespace=namespace, propagation_policy="Foreground"
-                )
-            except kubernetes.client.rest.ApiException:
-                pass
+        if should_cleanup:
+            for job in ["sayhi1", "sayhi2", "fail"]:
+                try:
+                    api_client.batch_api.delete_namespaced_job(
+                        job, namespace=namespace, propagation_policy="Foreground"
+                    )
+                except kubernetes.client.rest.ApiException:
+                    pass
